array iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,9 +11,9 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	long unsigned int i;
 
-	for (i = 0; i < (int) size; i++)
+	for (i = 0; i < size; i++)
 	{
 		action(array[i]);
 		_putchar('\n');
